requirements.txt
python-dotenv==0.21.0
python-frontmatter==1.0.0
requests==2.28.1
